third_party/blink/web_tests/platform/mac-mac10.11/external/wpt/largest-contentful-paint/observe-after-untrusted-scroll-expected.txt

This is a testharness.js-based test.
FAIL Same-origin image after a JS initiated scroll event is observable. assert_equals: expected 0 but got 31.715000000076543
Harness: the test ran to completion.
